+++ UI/Controls/LayersNameEditor.xaml.cs
@@ -28,7 +28,7 @@
         }
 
         private void LayersNameEditor_Loaded(object sender, System.Windows.RoutedEventArgs e) {
-            var layers = Enum.GetValues(typeof(Layers)).Cast<Layers>().OrderBy(x => (byte)x).ToList();
+            var layers = Enum.GetValues(typeof(Layers)).Cast<Layers>().OrderBy(x => (ushort)x).ToList();
             layers.Remove(Layers.None);
             layers.Remove(Layers.All);
